* Fixed bug in bsh writing * Fixed issue with GraphicsMagick not finding it's config files * Do the three "classic" lods form ovlcompiler xml files, not just one as ovlcompiler did
git-svn-id: https://svn.code.sf.net/p/rct3/code/trunk@76 68d9b628-aa21-0410-a705-e462ad5a1edf

+++ RCT3 Importer/src/libOVLng/ManagerBSH.cpp
@@ -264,7 +264,7 @@ unsigned char* ovlBSHManager::Make() {
             // Data transfer, indices
             c_mesh->Triangles = reinterpret_cast<unsigned short*>(c_data);
             c_data += c_mesh->IndexCount * sizeof(unsigned short);
-            memcpy(c_mesh->Triangles, m_modellist[i]->sh[s]->Triangles, c_mesh->IndexCount * sizeof(unsigned long));
+            memcpy(c_mesh->Triangles, m_modellist[i]->sh[s]->Triangles, c_mesh->IndexCount * sizeof(unsigned short));
             m_relman->AddRelocation((unsigned long *)&c_mesh->Triangles);
             DUMP_RELOCATION("ovlBSHManager::Make, Triangles", c_mesh->Triangles);
